styling dashboard/breakdown __ Adding Contact Section
finishing up with styling of the dashboard/breakdown, added a contact section. Also added buttons to be quick links that are fixed to the left and/or right of the screen. Will finish up with the styling of the contact form and then proceed to work on the landscape mode